+++ client/admin-side/cms/src/components/UserTableRow.js
@@ -14,27 +14,49 @@ export default function UserTableRow({ user, i }) {
         width: "150px",
         height: "150px"
     }
-    const handleDelete = (e, id) => {
-        e.preventDefault();
-        dispatch(deleteUser(id))
-            .then((res) => {
-                Swal.fire(
-                    res.data.message,
-                    'You clicked the button!',
-                    'success'
-                )
-            })
-            .then((data) => {
-                dispatch(fetchUsers(data))
-            })
-            .catch((err) => {
-                Swal.fire({
-                    icon: 'error',
-                    title: 'Oops...',
-                    text: err.response.data.message,
-                })
-            })
+
+    const deleteHandler = (id) => {
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            background: '#000',
+            color: '#fff',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                dispatch(deleteUser(id))
+                    .then(res => {
+                        Swal.fire({
+                            position: 'top-end',
+                            icon: 'success',
+                            background: '#000',
+                            color: '#fff',
+                            title: res.data.message,
+                            showConfirmButton: true,
+                            timer: 1000
+                        })
+                            .then(data => {
+                                dispatch(fetchUsers())
+                            })
+                    }).catch(err => {
+                        Swal.fire({
+                            icon: 'error',
+                            title: 'Oops...',
+                            text: err.response.data.message,
+                        })
+                    })
+            }
+
+        })
+
+
     }
+
+
 
     const handleEdit = (id) => {
         navigate(`/user/${id}`)
@@ -51,7 +73,7 @@ export default function UserTableRow({ user, i }) {
             <td>
                 <Stack gap={2} direction="horizontal">
                     <Button variant="primary" className="justify-center" onClick={() => handleEdit(user.id)} >Edit</Button>
-                    <Button variant="danger" className="justify-center" onClick={(e) => handleDelete(e, user.id)}>Delete</Button>
+                    <Button variant="danger" className="justify-center" onClick={() => deleteHandler(user.id)}>Delete</Button>
                 </Stack>
             </td>
         </tr>
